+++ FluentFlyoutWPF/Pages/HomePage.xaml
@@ -276,19 +276,26 @@
             </ui:Button>
         </StackPanel>
 
-        <TextBlock Text="{DynamicResource AboutPremiumTitle}" FontSize="14" FontWeight="SemiBold" Margin="0,36,0,0" />
-        <Grid Margin="0,12,0,0">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-            </Grid.RowDefinitions>
-
-            <ui:CardControl x:Name="UnlockFullExperienceTextCard" Tag="Indexable" Grid.Row="0">
+        <StackPanel Margin="0,36,0,0">
+            <StackPanel.Style>
+                <Style TargetType="StackPanel">
+                    <Setter Property="Visibility" Value="Collapsed" />
+                    <Style.Triggers>
+                        <MultiDataTrigger>
+                            <MultiDataTrigger.Conditions>
+                                <Condition Binding="{Binding IsStoreVersion}" Value="True" />
+                                <Condition Binding="{Binding IsPremiumUnlocked}" Value="False" />
+                            </MultiDataTrigger.Conditions>
+                            <Setter Property="Visibility" Value="Visible" />
+                        </MultiDataTrigger>
+                    </Style.Triggers>
+                </Style>
+            </StackPanel.Style>
+            <TextBlock Text="{DynamicResource AboutPremiumTitle}" FontSize="14" FontWeight="SemiBold" />
+            <ui:CardControl x:Name="UnlockFullExperienceTextCard" Margin="0,12,0,0">
                 <ui:CardControl.Header>
                     <StackPanel>
-                        <StackPanel Margin="0,0,0,12" Visibility="{Binding IsPremiumUnlocked, Converter={StaticResource BoolToCollapsedVisibleConverter}}">
+                        <StackPanel Margin="0,0,0,12">
                             <ui:TextBlock FontSize="14" Margin="0,0,0,8" TextWrapping="WrapWithOverflow"
                                   Text="{DynamicResource UnlockFullExperienceText}" />
                             <ui:TextBlock FontSize="14" Margin="12,0,12,4" Opacity="0.75" TextWrapping="WrapWithOverflow"
@@ -303,67 +310,16 @@
                                   Text="{Binding Source={StaticResource AllFuturePremiumFeatures}, StringFormat='•  {0}'}" />
                         </StackPanel>
 
-                        <!-- Premium Status InfoBar (only shown for Store versions) -->
-                        <ui:InfoBar Grid.Row="1" Title="{DynamicResource PremiumUnlockedMessage}" Severity="Success" IsClosable="True">
-                            <ui:InfoBar.Style>
-                                <Style TargetType="ui:InfoBar" BasedOn="{StaticResource {x:Type ui:InfoBar}}">
-                                    <Setter Property="IsOpen" Value="False" />
-                                    <Style.Triggers>
-                                        <MultiDataTrigger>
-                                            <MultiDataTrigger.Conditions>
-                                                <Condition Binding="{Binding IsStoreVersion}" Value="True"/>
-                                                <Condition Binding="{Binding IsPremiumUnlocked}" Value="True"/>
-                                            </MultiDataTrigger.Conditions>
-                                            <Setter Property="IsOpen" Value="True"/>
-                                        </MultiDataTrigger>
-                                    </Style.Triggers>
-                                </Style>
-                            </ui:InfoBar.Style>
-                        </ui:InfoBar>
-
-                        <!-- Premium Locked InfoBar (only shown for Store versions without premium) -->
-                        <ui:InfoBar Grid.Row="1" IsOpen="True" Severity="Warning" IsClosable="False" Margin="0,4,0,0">
-                            <ui:InfoBar.Style>
-                                <Style TargetType="ui:InfoBar">
-                                    <Setter Property="Visibility" Value="Collapsed"/>
-                                    <Style.Triggers>
-                                        <MultiDataTrigger>
-                                            <MultiDataTrigger.Conditions>
-                                                <Condition Binding="{Binding IsStoreVersion}" Value="True"/>
-                                                <Condition Binding="{Binding IsPremiumUnlocked}" Value="False"/>
-                                            </MultiDataTrigger.Conditions>
-                                            <Setter Property="Visibility" Value="Visible"/>
-                                        </MultiDataTrigger>
-                                    </Style.Triggers>
-                                </Style>
-                            </ui:InfoBar.Style>
-                            <StackPanel Orientation="Horizontal">
-                                <ui:Button Name="UnlockPremiumButton"
-                                    Click="UnlockPremiumButton_Click" 
-                                    Appearance="Primary"
-                                    Margin="4,0,0,8" FontSize="12">
-                                    <ui:Button.Content>
-                                        <StackPanel Orientation="Horizontal">
-                                            <TextBlock Text="{DynamicResource UnlockPremiumButton}" />
-                                            <TextBlock Text=" - " />
-                                            <TextBlock Text="{Binding PremiumPrice}" />
-                                        </StackPanel>
-                                    </ui:Button.Content>
-                                </ui:Button>
-                            </StackPanel>
-                        </ui:InfoBar>
-
-                        <!-- GitHub Version InfoBar -->
-                        <ui:InfoBar Grid.Row="1" Title="{DynamicResource PremiumGitHubMessage}" IsOpen="True" Severity="Informational" IsClosable="False" Visibility="{Binding IsStoreVersion, Converter={StaticResource BoolToCollapsedVisibleConverter}}" />
+                        <fluentflyoutcontrols:PremiumStatus />
                     </StackPanel>
                 </ui:CardControl.Header>
             </ui:CardControl>
+        </StackPanel>
 
-            <StackPanel Grid.Row="2" Orientation="Horizontal" Margin="0,12,0,0">
-                <ui:Anchor Content="{DynamicResource SponsorMeLink}" Icon="{ui:SymbolIcon Heart24}" NavigateUri="https://github.com/sponsors/unchihugo" Margin="0,0,4,0" />
-                <ui:Anchor Content="{DynamicResource BuyMeCoffeeLink}" Icon="{ui:SymbolIcon DrinkCoffee24}" NavigateUri="https://ko-fi.com/unchihugo" Margin="0,0,4,0" />
-            </StackPanel>
-        </Grid>
+        <StackPanel Orientation="Horizontal" Margin="0,24,0,0">
+            <ui:Anchor Content="{DynamicResource SponsorMeLink}" Icon="{ui:SymbolIcon Heart24}" NavigateUri="https://github.com/sponsors/unchihugo" Margin="0,0,4,0" />
+            <ui:Anchor Content="{DynamicResource BuyMeCoffeeLink}" Icon="{ui:SymbolIcon DrinkCoffee24}" NavigateUri="https://ko-fi.com/unchihugo" Margin="0,0,4,0" />
+        </StackPanel>
 
         <DockPanel LastChildFill="False" Margin="0,24,0,12">
             <StackPanel Orientation="Horizontal" DockPanel.Dock="Bottom" HorizontalAlignment="Left">
