+++ FluentFlyoutWPF/Windows/OnboardingWindow.xaml
@@ -6,6 +6,7 @@
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:ui="http://schemas.lepo.co/wpfui/2022/xaml"
     xmlns:controls="clr-namespace:MicaWPF.Controls;assembly=MicaWPF"
+    xmlns:fluentflyoutcontrols="clr-namespace:FluentFlyoutWPF.Controls"
     xmlns:viewModels="clr-namespace:FluentFlyoutWPF.ViewModels"
     Title="Welcome to FluentFlyout"
 	Icon="/Resources/FluentFlyout2.ico"
@@ -160,58 +161,7 @@
                                     <ui:TextBlock FontSize="14" Margin="12,0,12,0" Opacity="0.7" TextWrapping="WrapWithOverflow"
                                         Text="{Binding Source={StaticResource AllFuturePremiumFeatures}, StringFormat='•  {0}'}" />
 
-                                    <!-- Premium Status InfoBar (only shown for Store versions) -->
-                                    <ui:InfoBar Grid.Row="1" Title="{DynamicResource PremiumUnlockedMessage}" Severity="Success" IsClosable="True" Margin="0,8,0,0">
-                                        <ui:InfoBar.Style>
-                                            <Style TargetType="ui:InfoBar" BasedOn="{StaticResource {x:Type ui:InfoBar}}">
-                                                <Setter Property="IsOpen" Value="False" />
-                                                <Style.Triggers>
-                                                    <MultiDataTrigger>
-                                                        <MultiDataTrigger.Conditions>
-                                                            <Condition Binding="{Binding IsStoreVersion}" Value="True"/>
-                                                            <Condition Binding="{Binding IsPremiumUnlocked}" Value="True"/>
-                                                        </MultiDataTrigger.Conditions>
-                                                        <Setter Property="IsOpen" Value="True"/>
-                                                    </MultiDataTrigger>
-                                                </Style.Triggers>
-                                            </Style>
-                                        </ui:InfoBar.Style>
-                                    </ui:InfoBar>
-                            
-                                    <!-- Premium Locked InfoBar (only shown for Store versions without premium) -->
-                                    <ui:InfoBar Grid.Row="1" IsOpen="True" Severity="Warning" IsClosable="False" Margin="0,8,0,0">
-                                        <ui:InfoBar.Style>
-                                            <Style TargetType="ui:InfoBar">
-                                                <Setter Property="Visibility" Value="Collapsed"/>
-                                                <Style.Triggers>
-                                                    <MultiDataTrigger>
-                                                        <MultiDataTrigger.Conditions>
-                                                            <Condition Binding="{Binding IsStoreVersion}" Value="True"/>
-                                                            <Condition Binding="{Binding IsPremiumUnlocked}" Value="False"/>
-                                                        </MultiDataTrigger.Conditions>
-                                                        <Setter Property="Visibility" Value="Visible"/>
-                                                    </MultiDataTrigger>
-                                                </Style.Triggers>
-                                            </Style>
-                                        </ui:InfoBar.Style>
-                                        <StackPanel Orientation="Horizontal">
-                                            <ui:Button Name="UnlockPremiumButton"
-                                            Click="UnlockPremiumButton_Click" 
-                                            Appearance="Primary"
-                                            Margin="4,0,0,8" FontSize="12">
-                                                <ui:Button.Content>
-                                                    <StackPanel Orientation="Horizontal">
-                                                        <TextBlock Text="{DynamicResource UnlockPremiumButton}" />
-                                                        <TextBlock Text=" - " />
-                                                        <TextBlock Text="{Binding PremiumPrice}" />
-                                                    </StackPanel>
-                                                </ui:Button.Content>
-                                            </ui:Button>
-                                        </StackPanel>
-                                    </ui:InfoBar>
-
-                                    <!-- GitHub Version InfoBar -->
-                                    <ui:InfoBar Grid.Row="1" Title="{DynamicResource PremiumGitHubMessage}" IsOpen="True" Severity="Informational" IsClosable="False" Visibility="{Binding IsStoreVersion, Converter={StaticResource BoolToCollapsedVisibleConverter}}" Margin="0,8,0,0" />
+                                    <fluentflyoutcontrols:PremiumStatus Margin="0,8,0,0" />
                                 </StackPanel>
                             </StackPanel>
                         </ui:PassiveScrollViewer>
