+++ FluentFlyoutWPF/Windows/OnboardingWindow.xaml.cs
@@ -166,10 +166,4 @@ public partial class OnboardingWindow : MicaWindow
         SettingsWindow.ShowInstance();
         Close();
     }
-
-    // same as in AboutPage.xaml.cs
-    private async void UnlockPremiumButton_Click(object sender, RoutedEventArgs e)
-    {
-        FluentFlyoutWPF.Classes.LicenseManager.UnlockPremium(sender);
-    }
 }
